scan-engine/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
python-owasp-zap-v2==0.0.24
